% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdata.R
\docType{data}
\name{hdata}
\alias{hdata}
\title{Head and neck cancer data}
\format{
A dataframe with 565 rows and 104 variables
\describe{
\item{ID}{ID of subjects}
\item{leftcensoring}{Initial censoring time}
\item{death}{Survival event}
\item{death2}{death due to other causes}
\item{os}{Duration of overall survival}
\item{PFS}{Duration of progression free survival}
\item{Prog}{Progression event}
\item{GJB1,...,HMGCS2}{High dimensional covariates}}
}
\usage{
data(hdata)
}
\description{
High dimensional head and neck cancer gene expression data
}
\examples{
data(hdata)
}
\keyword{datasets}
