% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglwbysu.R
\name{rglwbysu}
\alias{rglwbysu}
\title{Bayesian univariate analysis of AFT model for selected covariates using regularization method.}
\usage{
rglwbysu(m, n, STime, Event, nc, ni, alpha, data)
}
\arguments{
\item{m}{Starting column number of covariates of study from high dimensional entered data.}

\item{n}{Ending column number of covariates of study from high dimensional entered data.}

\item{STime}{name of survival time in data}

\item{Event}{name of event in data. 0 is for censored and 1 for occurrence of event.}

\item{nc}{number of chain used in model.}

\item{ni}{number of iteration used in model.}

\item{alpha}{It is chosen value between 0 and 1 to know the regularization method. alpha=1 for Lasso,
alpha=0 for Ridge and alpha between 0 and 1 for elastic net regularization.}

\item{data}{High dimensional gene expression data that contains event status, survival time and and set of covariates.}
}
\value{
posterior estimates (Coef, SD, Credible Interval) of regression coefficient for all selected covariate (using regularization) in model and deviance.
}
\description{
Provides posterior Estimates of selected variable(by regularization technique) in AFT model for univariate in high dimensional gene expression data with MCMC. Incorporated variable selection has been done with regularization technique. It also deals covariates with missing values.
}
\details{
Here weibull distribution has been used for AFT model with MCMC.
This function deals covariates (in data) with missing values. Missing value in any column (covariate) is replaced by mean of that particular covariate.
}
\examples{
##
data(hdata)
set.seed(1000)
rglwbysu(9,45,STime="os",Event="death",2,10,1,hdata)
##

}
\references{
Prabhash et al(2016) <doi:10.21307/stattrans-2016-046>
}
\seealso{
wbysuni,wbysmv, rglwbysm
}
\author{
Atanu Bhattacharjee, Gajendra Kumar Vishwakarma and Pragya Kumari
}
